update cached turnitin data when redirecting to originality report
Turnitin has no way to register callbacks when a submission's
originality score changes.  However, teachers are able to
modify originality scores for submissions by marking certain
sources as being ignored when checking for originality.  If a
teacher does this then the data we've cached will get out of
date.

We don't really want to have to regularly poll for updated
info, so this will just schedule a background job to check
for updates whenever someone goes to view the full
originality report.  There will still be plenty of time
when the data it out of sync, but it at least gives a
way in the UI to update the data.

Change-Id: Ia8cc814a2e9cf5ec140d3fd22acedfbb00da1b8a
Reviewed-on: https://gerrit.instructure.com/2163
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>