make datetime_field() idempotent
There are some places where we might call datetime_field() on a field twice.
That causes the field to double its date when it's submitted. We can also just
be careful when calling datetime_field(), but I'm not sure there's a valid use
case for calling it twice, given how much it jacks up the control.

fixes #3818

Change-Id: I79987486efeb5c77497183a41251a6990747fd52
Reviewed-on: https://gerrit.instructure.com/2291
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>
Reviewed-by: Brian Whitmer <brian@instructure.com>
Tested-by: Zach Wily <zach@instructure.com>