fix display of gradebook in firefox if you zoom out.
the root of the problem is when we set the width of the
$row for the header and each row in the table.  each of
the .cell's are float:left'ed and so we need to set a
total width of the container.  when you zoom out the
border width of 1px did not shrink so there was no room
in the container to float:left the last cell so it
wrapped to a new line.

Change-Id: I01d3b4e04edc3befff3e4366a5b7f5afe5672703
fixes: #3891
Reviewed-on: https://gerrit.instructure.com/2409
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Whitmer <brian@instructure.com>