consolidate course permissions cache, use in grants_right
We have code to cache the course permissions for a user in the
permissions code itself, and also in Course#enrollment_allows. The
latter was also building up the cache one permission at a time, and
writing it back to memcache every time, causing a lot of needless
memcache writes.

grants_right?() was also skipping this cache since it doesn't call
grants_rights, so special case that as well.

Change-Id: I89ff5d44a0f50985e2ecfdefb68df869e7d68a1a
Reviewed-on: https://gerrit.instructure.com/2841
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Whitmer <brian@instructure.com>