let teacher always update quiz submissions; fixes #4176
Previously we were trying not to let a teacher update a quiz submission that
was in progress. However, this had the side effect of not allowing them to
ever update a quiz submission for a quiz that was being taken - even an older
quiz submission. We're removing that restriction. Normally teachers will only
see quiz submission versions that have been taken, so this won't be a problem.
However it means that if a teacher wants to be mean they can set a grade on a
quiz that's currently being taken by some form hacking. This seems like a silly
thing to prevent since a teacher can do whatever they want to student's grades
anyway.

Change-Id: I671985703089166e7aa06c57be62507c8171454e
Reviewed-on: https://gerrit.instructure.com/2955
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Hudson <hudson@instructure.com>