background scribd resubmission
If a scribd submission errors out, we add a delay
and make it a background task to resubmit the
scribd doc.  This is important since sometimes the
resubmission gets triggered as part of a page
request, and it also waits a little bit before
resubmitting to scribd in case the problem
is something to do with their servers.

Change-Id: I14e796b90bad371c8c2718fc83cd22c97d94cc96
Reviewed-on: https://gerrit.instructure.com/3026
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>