make sure delegated auth gets a stab at first-login. fixes #4773
URLs that might not require a user were using
render_unauthorized_action, which just rendered the login partial,
rather than redirecting to login_url.  This meant that delegated auth
logic in PseudoSessionsController#new was getting skipped.  So for
render_unauthorized_action in the specific case that there is no user,
call a method to kick of delegated auth, if necessary.

Also fix a problem that session[:return_to] was getting wiped for
delegated auth, so you'd end up at the dashboard, rather than where
you wanted to go.

Change-Id: I42c24e31ca58cccc33d349bcb451defd3aaa5c9c
Reviewed-on: https://gerrit.instructure.com/3993
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>