fix inconsistencies with preservation of 0-pts-possible pass/fail assignments
fixes #4932

Basically, we had an if in there so that 0-pts-possible assignments wouldn't
munge the grade, since score will always be 0.0 for pass *and* fail.  This
broke other assignments with 0-pts-possible (i.e. extra credit assignmens)
to never update their grade when only the score changed (i.e. a quiz updates
the score, but not the grade).  So, only do the special casing for pass/fail
assignments, as well as cleaning it up so that nil vs. 0 pts possible behaves
the same, and not putting in nil as the score for a nil-pts-possible pass
pass/fail assignment.

Change-Id: I6e7c74a685621db794a5a3b252578435bab5b27d
Reviewed-on: https://gerrit.instructure.com/4356
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>