avoid requesting all google docs just to see if we still have access
With some users that have thousands of files, that request can take
upwards of 10-15 seconds; even doing a head and/or limiting results
doesn't speed it up (silly googles).  So instead, do a head for the
api that returns info on the token (yes, it says it's for AuthSub,
but it works for OAuth as well).

Change-Id: If838d0f6f04f313887dc734786d5cdeffa3e28a5
Reviewed-on: https://gerrit.instructure.com/4561
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>