can't create graded quiz from assignments more options page
There's a problem where if you start to create a quiz from the
assignments list page and click "more options", you won't be able
to actually save the assignment. The problem is that the form
validation is failing, because the "Submission:" option is set to
"Online Submission" but no submission types are selected.
However, since this option is hidden on quiz type assignments,
the validation error bubble never appears and the form just
won't submit.

fixes #5023

Change-Id: Iecb843980b7e1c56fff313331bd30d2a79eee9d8
Reviewed-on: https://gerrit.instructure.com/4613
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Tested-by: Hudson <hudson@instructure.com>