return secondary_identifier fields if can manage_students
Before, in the api, we only returned sis_user_id
and sis_login_id if they were an account admin.

This relaxes the restriction to also include anyone
that can manage_students in the course.  It also
returns a login_id, which is the unique_id column
from the database.

Change-Id: Ice36c5414b48a706c10d337533778a78da3b3f5e
Reviewed-on: https://gerrit.instructure.com/4648
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Hudson <hudson@instructure.com>