speed up courses selenium test
In this spec we call find_element a lot when we expect that the element
probably isn't there. implicit_wait was causing each of those calls to take 3
seconds. For this spec, we manually set it to zero.

Also changed the default 3 second implicit_wait to 1 second for everything.

Also changed the spec to use create_session instead of doing a full login
sequence.

These changes reduced this spec's runtime from 90s to 30.

Change-Id: I9f268ac437eae5ef24ec71a5c1108957b83fd8de
Reviewed-on: https://gerrit.instructure.com/4794
Tested-by: Hudson <hudson@instructure.com>
Tested-by: Selenium <selenium@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>