fix more selenium tests
The wait_for_ajax at the end of every test was failing on pages that didn't
have a Canvas page loaded at the end of the test.

Fixed a bug that was causing us to wait longer than necessary for the forked
servers to start up.

Also did some other cleanup.

Change-Id: Ia8a2cdc2dcb2967427a698ff200ed75485a495ef
Reviewed-on: https://gerrit.instructure.com/5397
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Bryan Madsen <bryan@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>