option to open BLTI links in a new tab
We are getting requests from some vendors to have some way to
load tools in a new tab. The preferred behavior is still to load
tools in the current window, but if a new tab is absolutely
necessary then there is the option now.

We can't reliably open links in a new tab without user intervention,
so we provide a button that users can click to load the tool
themselves in a new tab. That means we also need to add some
timeout functionality in case they take more than a few minutes
to load the tool, since BLTI links expire over time.

fixes #5814

Change-Id: I126f1aaa9cdcfe49b8f2b8ef8b7fbbaf8c56fe67
Reviewed-on: https://gerrit.instructure.com/5957
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>