comment out not-yet-applicable-spec refs #5833
This spec was written *after* a commit that is not yet submitted,
but it was submitted without realizing the dependency on that
commit.  So just comment out the spec now (the code change it
went with is still applicable to current master, just not testable
until the other commit is submitted that allows more direct
creation of the necessary conditions).

Change-Id: Ibc615b3d77bf9757abdfa4b41c8752b4a8611cbe
Reviewed-on: https://gerrit.instructure.com/6364
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>