support creation_pending users that were created prior to the big refactor
refs #5833

prior to the big refactor, creation_pending users were created with
pseudonyms.  after the refactor they aren't (since there could be
conflicts that need to be resolved by the user).  without this fix,
the old users that have a pseudonym would be forced to create another,
non-conflicting pseudonym.  this fix just reuses the pre-created
pseudonym, if it exists

Change-Id: Ib3dc822e6ca179fb0e512d8bc4b3014cdaad262a
Reviewed-on: https://gerrit.instructure.com/6483
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>