placeholders for quiz answers fixes #5438
The previous fix for this in #4949 broke some
other question types where the app would get stuck
in a loop of selecting 3-4 inputs rendering the
page useless.

Added jQuery.fn.placeholder, can now use the
placeholder attribute on input elements, and then
call $(selector).placeholder(); to support IE.

The plugin checks for support and does nothing for
browsers that support placeholder, so we don't
need to do it ourselves.

also fixes #4949 (again)

Change-Id: Ifbb3e8110ea259ccbe235e3e364050ab29a51b68
Reviewed-on: https://gerrit.instructure.com/6572
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>
Tested-by: Jon Jensen <jon@instructure.com>