don't show discussion responses in stream when not allowed
Discussions have an option to not allow users to see replies
until they've posted themselves. This worked correctly on the
discussion page itself, but the user would still see the
replies in their stream

This also removes that option, and the delayed posting option
from group discussion topics. These settings are useless in
that context because all users are admins and can see the
topics anyway.

Test Plan:
 * create a topic in a course and check the box "Replies are not visible until after users post"
 * post a reply to that topic as the teacher
 * log in as a student in that course and go to your dashboard
 * you should see the topic in the stream but not the response you created as a teacher

closes #4706

Change-Id: I0c50a3adb8e18ea74d209f4c9a5ac429b7faaf0e
Reviewed-on: https://gerrit.instructure.com/6684
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>