allow forcing the server host and ip to be used for selenium tests
This is useful when running selenium specs in an environment where you need to
specify an SSH-forwarded tunnel to communicate over, for example.

test plan: set server_ip and server_port in selenium.yml and run the selenium
tests. The remote selenium session should use those when accessing the test
host.

Change-Id: Iec5963a8c1481c25fa79820802c9b2ce35172f4a
Reviewed-on: https://gerrit.instructure.com/7248
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>