allow module progression when background job hasn't finished
If a user just completed a module and hasn't gone to the
modules page and clicks the next module item button that
goes into the next module (not another item within the same
module), the next module will still be locked. (If they click
before the delayed job is run)

This commit makes it so that clicking on a module item link
that goes to a locked item will cause the module progressions
to be recalculated in-request before the user is redirected
to the destination resource. This should happen fairly
infrequently since the user will only be able to access a
locked module item link under the circumstances mentioned
in the first paragraph. (locked items are disabled while
on the modules page)

Test Plan:
 * Run canvas in production with caching on and Delayed Jobs not running
 * Create a module with a quiz as the last item
 * Make it so that the module isn't complete unless the student has scored at least 1
 * Create a second module with an assignment as the first item
 * Make it so that module isn't unlocked until the first module is complete
 * Add a student and publish the course
 * Log in as student
 * Go to the quiz in the first module
 * Observe that clicking the next module item button takes you to the assignment but that it is locked.
 * Take the quiz and score at least 1, _do not_ go to the modules page after this is done.
 * Click the next module item button and make sure the assignment is unlocked

closes #6254

Change-Id: I0719ecdc75af2d5c61bf78fbe2f07c4f21f2e24d
Reviewed-on: https://gerrit.instructure.com/7362
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>