don't redirect to /courses/X xhr to settings
for a user with :read_as_admin but not :read permissions to a course
(odd, but currently happens for designer enrollments in an unpublished
course) the user is redirected from the course "Home" to the course
settings. however, the settings page turns around and makes an xhr
request to /courses/X again for some data. don't redirect that call or
you get an infinite loop.

also, this change cuts out the unnecessary remainder of
CoursesController#show for xhr requests; before we still did all that
work and then just ignored it.

Change-Id: I4b6d39c8bf4e12173a4aa168903d2148f89c7a3d
test-plan:
  - create an unpublished course
  - enroll a designer in that course (currently only possible via sis or
    console)
  - while logged in/masquerading as the designed, visit the course's
    settings page
  - confirm there is *not* a runaway xhr recursion
Reviewed-on: https://gerrit.instructure.com/8039
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Ryan Florence <ryanf@instructure.com>