fix page errors related to submission versions in speedgrader; fixes #6869
In the SpeedGrader, if you went past a user with > 1 submission, grading
subsequent students with only one submission would result in a page error in
the submission content iframe after the grade is saved. We were only
populating the <select> element for > 1 submissions people, and the change()
binding on it would call for everyone else, triggering badness.

test plan:
 * Create 2 users in a course with an assignment.
 * Submit the assignment twice as student A, and once as student B.
 * In SpeedGrader for that assignment, browser to student A, then student B.
 * Grade student B's submission.
 * Notice that when the submission reloads in the iframe, it is not a page
   error.

Change-Id: If461ecc1121fcae0549f1f451ecaa33edf2aa69e
Reviewed-on: https://gerrit.instructure.com/8046
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Tested-by: Hudson <hudson@instructure.com>